Refactored "asciiDraw" methods out of `Tile` and into extension methods.
